requirements.txt
streamlit
pinecone
pypdf
python-dotenv
google-generativeai
langchain